backend/requirements.txt
# Required for FastAPI test client
httpx>=0.24,<0.28
starlette>=0.36,<0.37
pytest
pytest==8.2.0
pytest_asyncio==0.23.6
fastapi>=0.110.0
pydantic>=2.0.0
# Required for MLflow and pkg_resources
setuptools
mlflow==2.12.1
cloudpickle==3.0.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pynvml==11.5.0
pyyaml==6.0.1
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0
tqdm==4.66.2
psutil==5.9.8
pynvml==11.5.0
python-multipart==0.0.9
hydra-core==1.3.2
pre-commit==3.7.0
uvicorn